backend/requirements.txt
# Django Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# CORS
django-cors-headers>=4.3,<5.0

# Filtering & Search
django-filter>=24.0

# Image handling
Pillow>=10.0

# PostgreSQL
psycopg2-binary>=2.9

# Environment variables
python-decouple>=3.8

# Development tools
django-extensions>=3.2
